# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SceneIt one-video ingestion and semantic-search proof.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/status:
    get:
      operationId: getBillingStatus
      tags: [billing]
      summary: Read the current account's membership and usage
      description: Returns private, owner-scoped billing state. The response must not be cached.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Current membership and, when active, the effective monthly usage window
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingStatus" }
        "401": { $ref: "#/components/responses/BillingFailure" }
        "503": { $ref: "#/components/responses/BillingFailure" }
  /billing/checkout:
    post:
      operationId: createBillingCheckout
      tags: [billing]
      summary: Create an allowlisted Stripe-hosted subscription checkout
      description: Requires pilot admission. Price, customer, ownership, and return destinations are derived by the server.
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillingCheckout" }
      responses:
        "200":
          description: Short-lived hosted checkout destination
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingHostedSession" }
        "400": { $ref: "#/components/responses/BillingFailure" }
        "401": { $ref: "#/components/responses/BillingFailure" }
        "402": { $ref: "#/components/responses/BillingFailure" }
        "403": { $ref: "#/components/responses/BillingFailure" }
        "409": { $ref: "#/components/responses/BillingFailure" }
        "429": { $ref: "#/components/responses/BillingFailure" }
        "500": { $ref: "#/components/responses/BillingFailure" }
        "503": { $ref: "#/components/responses/BillingFailure" }
  /billing/portal:
    post:
      operationId: createBillingPortal
      tags: [billing]
      summary: Create an owner-scoped Stripe-hosted customer portal
      description: Remains available to an authenticated billing customer without pilot admission or active membership.
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmptyRequest" }
      responses:
        "200":
          description: Short-lived hosted customer-portal destination
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingHostedSession" }
        "400": { $ref: "#/components/responses/BillingFailure" }
        "401": { $ref: "#/components/responses/BillingFailure" }
        "403": { $ref: "#/components/responses/BillingFailure" }
        "404": { $ref: "#/components/responses/BillingFailure" }
        "409": { $ref: "#/components/responses/BillingFailure" }
        "429": { $ref: "#/components/responses/BillingFailure" }
        "500": { $ref: "#/components/responses/BillingFailure" }
        "503": { $ref: "#/components/responses/BillingFailure" }
  /billing/webhook:
    post:
      operationId: receiveBillingWebhook
      tags: [billing]
      summary: Verify and durably process a Stripe webhook
      description: The server validates Stripe-Signature against the unmodified bounded request body. This is not a browser endpoint.
      security: [{ stripeSignature: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillingWebhookEvent" }
      responses:
        "200":
          description: Processing completed or was durably recorded for bounded recovery
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingWebhookAck" }
        "400": { $ref: "#/components/responses/BillingFailure" }
        "409": { $ref: "#/components/responses/BillingFailure" }
        "413": { $ref: "#/components/responses/BillingFailure" }
        "500": { $ref: "#/components/responses/BillingFailure" }
        "503": { $ref: "#/components/responses/BillingFailure" }
  /login:
    get:
      operationId: login
      parameters:
        - { name: returnTo, in: query, schema: { type: string } }
      responses:
        "302": { description: OIDC authorization redirect }
...
  /logout:
    post:
      operationId: logout
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
...
  /imports:
    post:
      operationId: createImport
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateImport"
      responses:
        "200": { $ref: "#/components/responses/ImportResponse" }
        "400": { $ref: "#/components/responses/Failure" }
        "401": { $ref: "#/components/responses/Failure" }
        "402": { $ref: "#/components/responses/BillingFailure" }
        "409": { $ref: "#/components/responses/Failure" }
        "429": { $ref: "#/components/responses/BillingFailure" }
        "503": { $ref: "#/components/responses/BillingFailure" }
...
  /proof:
    get:
      operationId: getProof
      tags: [proof]
      summary: Read the current one-video proof and its measured checks
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Current proof
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proof"
        "401": { $ref: "#/components/responses/ProtectedFailure" }
        "403": { $ref: "#/components/responses/ProtectedFailure" }
        "503": { $ref: "#/components/responses/ProtectedFailure" }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: __Host-sceneit_session }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
    stripeSignature: { type: apiKey, in: header, name: Stripe-Signature }
...
  responses:
    ImportResponse:
      description: Current private import
      content:
        application/json:
          schema: { $ref: "#/components/schemas/VideoImport" }
    Failure:
      description: Safe, actionable failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    ProtectedFailure:
      description: Typed protected-resource failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtectedApiFailure"
    BillingFailure:
      description: Safe billing, entitlement, quota, capacity, or validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BillingProblem"
  schemas:
    EmptyRequest:
      type: object
      additionalProperties: false
...
    BillingCheckout:
      type: object
      additionalProperties: false
      required: [plan, idempotencyKey]
      properties:
        plan: { type: string, enum: [monthly, yearly] }
        idempotencyKey: { type: string, format: uuid }
    BillingHostedSession:
      type: object
      additionalProperties: false
      required: [url, expiresAt]
      properties:
        url: { type: string, format: uri }
        expiresAt: { type: [string, "null"], format: date-time }
    BillingWebhookEvent:
      type: object
      description: Provider event whose exact raw bytes are verified before parsing.
      additionalProperties: true
    BillingWebhookAck:
      type: object
      additionalProperties: false
      required: [received]
      properties:
        received: { type: boolean, const: true }
    BillingMetric:
      type: object
      additionalProperties: false
      required: [limit, used, remaining]
      properties:
        limit: { type: integer, minimum: 0 }
        used: { type: integer, minimum: 0 }
        remaining: { type: integer, minimum: 0 }
    BillingMetrics:
      type: object
      additionalProperties: false
      required: [imports, upload_attempts, analysis_seconds, searches, media_bytes, frames]
      properties:
        imports: { $ref: "#/components/schemas/BillingMetric" }
        upload_attempts: { $ref: "#/components/schemas/BillingMetric" }
        analysis_seconds: { $ref: "#/components/schemas/BillingMetric" }
        searches: { $ref: "#/components/schemas/BillingMetric" }
        media_bytes: { $ref: "#/components/schemas/BillingMetric" }
        frames: { $ref: "#/components/schemas/BillingMetric" }
    BillingUsage:
      type: object
      additionalProperties: false
      required: [windowStart, windowEnd, metrics, storage, workStopped]
      properties:
        windowStart: { type: [string, "null"], format: date-time }
        windowEnd: { type: [string, "null"], format: date-time }
        metrics: { $ref: "#/components/schemas/BillingMetrics" }
        storage: { $ref: "#/components/schemas/BillingMetric" }
        workStopped: { type: boolean }
    BillingStatus:
      type: object
      additionalProperties: false
      required: [enabled, environment, membership, paidThrough, cancelAtPeriodEnd, usage]
      properties:
        enabled: { type: boolean }
        environment: { type: [string, "null"], enum: [test, live, null] }
        membership: { type: string, enum: [disabled, active, inactive] }
        paidThrough: { type: [string, "null"], format: date-time }
        cancelAtPeriodEnd: { type: boolean }
        usage:
          anyOf:
            - $ref: "#/components/schemas/BillingUsage"
            - { type: "null" }
    BillingProblem:
      type: object
      additionalProperties: false
      description: |
        Safe commercial-work failure. Current domain code/status mappings are:
        `invalid_request`, `invalid_plan`, `invalid_idempotency_key`, `invalid_webhook` (400);
        `http_401` (401); `membership_required` (402);
        `pilot_not_admitted`, `origin_rejected`, `http_403` (403);
        `billing_customer_missing` (404);
        `idempotency_conflict`, `checkout_outcome_unknown`, `subscription_exists`,
        `billing_environment_mismatch`, `billing_customer_outcome_unknown`,
        `billing_relationship_invalid`, `reservation_conflict`, `reservation_released`,
        `webhook_rejected`, `invoice_not_paid`, `checkout_completed`, `checkout_exists`,
        `checkout_plan_conflict`, `checkout_state_invalid` (409);
        `participant_throttled`, `owner_quota_exhausted`, `storage_quota_exhausted` (429);
        and `billing_disabled`, `billing_unavailable`, `billing_provider_unavailable`,
        `provider_outcome_unknown`, `provider_rejected`, `service_work_stopped`,
        `service_capacity_exhausted`, `database_unavailable`,
        `database_capacity_exhausted`, `upstream_unavailable` (503).
        `webhook_too_large` is 413 and an unexpected failure is `internal_error` (500).
        Validation failures use `invalid_request`; `validation_error` is reserved for
        compatible validation emitters.
      required: [error, code, state]
      properties:
        error: { type: string }
        code:
          type: string
          enum:
            - membership_required
            - billing_disabled
            - service_work_stopped
            - service_capacity_exhausted
            - owner_quota_exhausted
            - storage_quota_exhausted
            - billing_unavailable
            - validation_error
            - invalid_request
            - invalid_plan
            - invalid_idempotency_key
            - pilot_not_admitted
            - origin_rejected
            - http_401
            - http_403
            - billing_customer_missing
            - billing_environment_mismatch
            - billing_customer_outcome_unknown
            - billing_relationship_invalid
            - idempotency_conflict
            - checkout_outcome_unknown
            - checkout_completed
            - checkout_exists
            - checkout_plan_conflict
            - checkout_state_invalid
            - subscription_exists
            - reservation_conflict
            - reservation_released
            - provider_outcome_unknown
            - provider_rejected
            - invalid_webhook
            - webhook_too_large
            - webhook_rejected
            - invoice_not_paid
            - billing_provider_unavailable
            - participant_throttled
            - database_unavailable
            - database_capacity_exhausted
            - upstream_unavailable
            - internal_error
        state:
          type: string
          enum: [admission_required, quota_exhausted, processing, uncertain, service_unavailable, not_found, unauthorized]
        retryable: { type: boolean, default: false }
        retryAfterSeconds: { type: ["integer", "null"], minimum: 0 }
...
    CreateImport:
      type: object
      additionalProperties: false
      required: [entryMethod, analysisAuthorized, playbackAuthorized, idempotencyKey]
      properties:
        entryMethod: { type: string, enum: [upload, link] }
        sourceUrl: { type: string, maxLength: 2048 }
        analysisAuthorized: { type: boolean, enum: [true] }
        playbackAuthorized: { type: boolean }
        idempotencyKey: { type: string, format: uuid }
...
    VideoImport:
      type: object
      required: [id, title, entryMethod, sourceKind, sourceUrl, externalId, state, statusMessage, progressPercent, errorCode, durationSeconds, fileSizeBytes, hasAudio, sourcePlaybackAvailable, sourcePlaybackUrl, playbackAuthorized, timelineStatus, createdAt, updatedAt, expiresAt, searchesUsed, searchLimit, importsUsed, importLimit, quotaMode]
      properties:
        id: { type: string }
        title: { type: string }
        entryMethod: { type: string, enum: [upload, link] }
        sourceKind: { type: string, enum: [file, youtube, x, tiktok, vimeo] }
        sourceUrl: { type: [string, "null"] }
        externalId: { type: [string, "null"] }
        state: { type: string, enum: [file_required, awaiting_upload, queued, resolving, validating, uploading, processing, indexing, ready, failed, needs_review, cancel_requested, cancelled, expired] }
        statusMessage: { type: string }
        progressPercent: { type: [number, "null"] }
        errorCode: { type: [string, "null"] }
        durationSeconds: { type: [number, "null"] }
        fileSizeBytes: { type: [integer, "null"] }
        hasAudio: { type: [boolean, "null"] }
        sourcePlaybackAvailable: { type: boolean }
        sourcePlaybackUrl: { type: [string, "null"] }
        playbackAuthorized: { type: boolean }
        timelineStatus: { type: string, enum: [not_applicable, unverified] }
        createdAt: { type: string }
        updatedAt: { type: string }
        expiresAt: { type: string }
        searchesUsed: { type: integer }
        searchLimit: { type: integer }
        importsUsed: { type: integer }
        importLimit: { type: integer }
        quotaMode:
          type: string
          enum: [lifetime, monthly]
          description: Semantics of the reported import and search counters.
...
    ApiFailure:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
    ProtectedApiFailure:
      type: object
      additionalProperties: false
      required: [error, code, state]
      properties:
        error: { type: string }
        code: { type: string }
        state:
          type: string
          enum: [admission_required, quota_exhausted, processing, uncertain, service_unavailable, not_found, unauthorized]
        retryable: { type: boolean, default: false }
        retryAfterSeconds: { type: ["integer", "null"], minimum: 0 }
...
    ProofCheck:
      type: object
      required: [id, label, status, detail]
      properties:
        id: { type: string }
        label: { type: string }
        status: { type: string, enum: [passed, pending, failed, unverified] }
        detail: { type: string }
    Proof:
      type: object
      required: [id, title, youtubeVideoId, youtubeUrl, durationSeconds, width, height, fileSizeBytes, hasAudio, state, statusMessage, model, timelineStatus, checks, sourcePlaybackAvailable, sourcePlaybackUrl, searchesUsed, searchLimit, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        youtubeVideoId: { type: string }
        youtubeUrl: { type: string }
        durationSeconds: { type: number }
        width: { type: integer }
        height: { type: integer }
        fileSizeBytes: { type: integer }
        hasAudio: { type: boolean }
        state: { type: string, enum: [queued, uploading, processing, indexing, ready, failed, needs_review] }
        statusMessage: { type: string }
        model: { type: string }
        timelineStatus: { type: string, enum: [unverified, verified, mismatch] }
        sourcePlaybackAvailable: { type: boolean }
        sourcePlaybackUrl: { type: ["string", "null"] }
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ProofCheck"
        searchesUsed: { type: integer }
        searchLimit: { type: integer }
        updatedAt: { type: string }
...